requirements.txt
python-dotenv>=1.0.1
web3>=6.15.1
requests>=2.32.3
